Move: Op to slatekit.common.data
