+++ src/lib/kotlin/slatekit-data/src/main/kotlin/slatekit/data/features/Purgable.kt
@@ -2,7 +2,7 @@ package slatekit.data.slatekit.data.features
 
 import slatekit.common.DateTime
 import slatekit.data.features.Deletable
-import slatekit.query.Op
+import slatekit.common.data.Op
 
 interface Purgeable<TId, T> : Deletable<TId, T> where TId : Comparable<TId> {
     /**
