+++ src/lib/kotlin/slatekit-entities/src/main/kotlin/slatekit/entities/features/Deletes.kt
@@ -4,7 +4,7 @@ import kotlin.reflect.KProperty
 import slatekit.entities.Entity
 import slatekit.entities.core.EntityOps
 import slatekit.query.IQuery
-import slatekit.query.Op
+import slatekit.common.data.Op
 import slatekit.results.Try
 import slatekit.results.builders.Tries
