+++ src/lib/kotlin/slatekit-entities/src/main/kotlin/slatekit/entities/features/UniqueIds.kt
@@ -4,7 +4,7 @@ import java.util.*
 import slatekit.entities.Entity
 import slatekit.entities.EntityWithUUID
 import slatekit.entities.core.EntityOps
-import slatekit.query.Op
+import slatekit.common.data.Op
 
 /**
  * NOTE: There is only 1 type constraint on the type parameter T to be Entity
