+++ src/lib/kotlin/slatekit-query/src/main/kotlin/slatekit/query/IQuery.kt
@@ -13,6 +13,8 @@
 
 package slatekit.query
 
+import slatekit.common.data.Op
+
 /**
  * Created by kreddy on 12/24/2015.
  */
