+++ src/lib/kotlin/slatekit-query/src/main/kotlin/slatekit/query/Query.kt
@@ -13,6 +13,7 @@
 
 package slatekit.query
 
+import slatekit.common.data.Op
 import java.util.concurrent.atomic.AtomicInteger
 
 /**
